refactoring : convert 기능은 부모 클래스에게 위임 및 추상화된 기능 사용
